+++ src/PrestaShopBundle/Resources/config/services/adapter/order.yml
@@ -204,7 +204,7 @@ services:
         command: PrestaShop\PrestaShop\Core\Domain\Order\Command\CancelOrderProductCommand
 
   prestashop.adapter.presenter.order_link_presenter:
-    class: PrestaShop\PrestaShop\Core\Domain\Order\Presenter\OrderLinkPresenter
+    class: PrestaShop\PrestaShop\Adapter\Presenter\Order\OrderLinkPresenter
     arguments:
       - '@router.default'
